Renamed folder "functions" to "helpers" & Created ConsoleLogger class to reduce ConfigurationManager functionalities
