+++ src/commands/download-video.command.ts
@@ -2,6 +2,7 @@ import fs from "fs";
 import { ArgumentsCamelCase } from "yargs";
 import ConfigurationManager from "../core/configuration/ConfigurationManager";
 import YTDownloader from "../core/YTDownloader";
+import ConsoleLogger from "../helpers/ConsoleLogger";
 import { DownloadFlags } from "../types/interface";
 
 export function handleVideoDownloadCommand(argv: ArgumentsCamelCase<DownloadFlags>) {
@@ -9,7 +10,7 @@ export function handleVideoDownloadCommand(argv: ArgumentsCamelCase<DownloadFlag
   if (argv.dir && fs.existsSync(argv.dir)) {
     saveDir = argv.dir;
   } else if (argv.dir && !fs.existsSync(argv.dir)) {
-    ConfigurationManager.showError(`"${argv.dir}" is not a valid directory`);
+    ConsoleLogger.showError(`"${argv.dir}" is not a valid directory`);
   }
 
   YTDownloader.shouldGenerateLogs = ConfigurationManager.shouldGenerateLogs;
